requirements.txt
attrs==22.1.0
requests==2.28.1
types-requests==2.28.11.7
python-dateutil==2.8.2
types-python-dateutil==2.8.19.5
sqlite-integrated==0.0.4
pandas==1.5.2
matplotlib==3.6.2